app/requirements.txt
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
exceptiongroup==1.2.1
fonttools==4.53.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
joblib==1.4.2
kiwisolver==1.4.5
leia-br==0.0.1
matplotlib==3.9.0
nltk==3.9.1
numpy==2.0.0
ollama==0.2.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.9.11
requests==2.32.3
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
